Moved selu weight and bias functions to models utils.py
